% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\docType{data}
\name{phrases}
\alias{phrases}
\title{A List of Import Phrases}
\format{
An object of class \code{character} of length 23.
}
\usage{
phrases
}
\description{
A collection of phrases used by the \code{\link[=import]{import()}} function to announce
the imposition of tariffs on foreign \code{R} packages.
}
\keyword{internal}
