% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tariff-package.R
\docType{package}
\name{tariff-package}
\alias{tariff}
\alias{tariff-package}
\title{Make Importing Great Again}
\description{
\code{tariff} is a fantastic tool that lets you impose import tariffs on \code{R}
packages. We're going to bring manufacturing BACK to your codebase by making
foreign imports more EXPENSIVE! Foreign packages have been STEALING our CPU
cycles for TOO LONG! It's time to put AMERICA FIRST and make importing FAIR
and BALANCED again! No more free rides for lazy, unoptimized dependencies
from overseas. We're DONE being the global garbage collector. Under this new
tariff plan, every line of code works for YOU—the hardworking developer!
We're protecting OUR cycles, OUR memory, and OUR beautiful, beautiful build
times. Believe me—your imports will never take advantage of you again. We're
taking control, and we're WINNING like never before!
}
\details{
See the function \code{\link[=tariff]{tariff()}} to learn how to impose tariffs and brings
back billions and billions and billions of CPU cycles to your codebase.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tariff.mihaiconstantin.com}
  \item Report bugs at \url{https://github.com/mihaiconstantin/tariff/issues}
}

}
\author{
\strong{Maintainer}: Mihai Constantin \email{mihai@mihaiconstantin.com} (\href{https://orcid.org/0000-0002-6460-0107}{ORCID}) (https://mihaiconstantin.com)

Authors:
\itemize{
  \item Huan Xu \email{huan.xu71@gmail.com} (https://github.com/hxu296)
}

}
\keyword{internal}
